refactor(web): make SizeFields size-solving injectable
SizeFields previously solved automatic sizes using partition-specific hooks
(route collection/index, partitionable lookups), which prevented reuse by the
logical volume form. Lift the size-solving behind a useSolvedSizes prop so
each host form supplies its own. The partition form keeps its solving in a
dedicated use-solved-sizes hook; the shared SizeFields stays form-agnostic.
